完善主题设置，迁移阅读器设置项
