+++ app/build.gradle.kts
@@ -19,7 +19,7 @@ android {
         minSdk = 24
         targetSdk = 35
         // 版本号为x.y.z则versionCode为x*1000000+y*10000+z*1000+debug版本号(开发需要时迭代, 三位数)
-        versionCode = 1_01_00_024
+        versionCode = 1_01_00_025
         versionName = "1.1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
