+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/components/SettingsEntry.kt
@@ -194,6 +194,8 @@ private fun SettingsSliderEntry(
             .clip(RoundedCornerShape(6.dp))
             .background(MaterialTheme.colorScheme.surfaceContainerHigh)
             .wrapContentHeight()
+            .then(modifier)
+            .wrapContentHeight()
             .padding(start = 18.dp, end = 14.dp)
             .padding(vertical = 8.dp),
         verticalAlignment = Alignment.CenterVertically,
