+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/ui/home/settings/theme/ThemeNavigation.kt
@@ -5,6 +5,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import indi.dmzz_yyhyy.lightnovelreader.data.userdata.UserDataPath
+import indi.dmzz_yyhyy.lightnovelreader.theme.LocalAppTheme
 import indi.dmzz_yyhyy.lightnovelreader.ui.LocalNavController
 import indi.dmzz_yyhyy.lightnovelreader.ui.book.reader.navigateToColorPickerDialog
 import indi.dmzz_yyhyy.lightnovelreader.ui.navigation.Route
@@ -24,15 +25,23 @@ fun NavGraphBuilder.settingsThemeDestination() {
         val navController = LocalNavController.current
         val viewModel = hiltViewModel<ThemeViewModel>()
         val readerSettingState = viewModel.settingState
+        val isDark = LocalAppTheme.current.isDark
         ThemeScreen(
             themeSettingState = readerSettingState,
             onClickBack = navController::popBackStackIfResumed,
             onClickChangeTextColor = {
                 navController.navigateToColorPickerDialog(
-                    UserDataPath.Reader.TextColor.path,
+                    if (isDark) UserDataPath.Reader.TextDarkColor.path
+                    else UserDataPath.Reader.TextColor.path,
                     listOf(-1, 0xFF1D1B20, 0xFFE6E0E9)
                 )
-            }
+            },
+            onClickChangeBackgroundColor = {
+                navController.navigateToColorPickerDialog(
+                    UserDataPath.Reader.BackgroundColor.path,
+                    listOf(-1, 0x38E8CCA5, 0x38FF8080, 0x38d3b17d,0x3834C759, 0x3832ADE6, 0x38007AFF, 0x385856D6, 0x38AF52DE)
+                )
+            },
         )
     }
 }
