+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/utils/Launcher.kt
@@ -23,3 +23,24 @@ fun uriLauncher(block: (Uri) -> Unit): ManagedActivityResultLauncher<Intent, Act
         }
     }
 }
+
+@Composable
+fun uriLauncherWithFlag(
+    block: (Uri, Boolean) -> Unit
+): Pair<ManagedActivityResultLauncher<Intent, ActivityResult>, (Boolean) -> Unit> {
+    var isDarkFlag by remember { mutableStateOf(false) }
+
+    val launcher = rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { activityResult ->
+        if (activityResult.resultCode == Activity.RESULT_OK) {
+            activityResult.data?.data?.let { uri ->
+                block(uri, isDarkFlag)
+            }
+        }
+    }
+
+    val setFlag: (Boolean) -> Unit = { flag ->
+        isDarkFlag = flag
+    }
+
+    return launcher to setFlag
+}
